+++ fuzz/run.sh
@@ -56,7 +56,8 @@ readonly BASE_TARGETS=(fuzz_scan fuzz_ac3_decode fuzz_eac3_decode fuzz_wav_read
                        fuzz_iec61937_unwrap fuzz_emdf_parse fuzz_oamd_parse
                        fuzz_joc_parse fuzz_osc_parse fuzz_signing_verify fuzz_matroska_demux
                        fuzz_mp4_demux fuzz_mpegts_demux fuzz_iab_parse fuzz_ac4_parse
-                       fuzz_sendspin_json fuzz_sendspin_frames fuzz_sendspin_handshake)
+                       fuzz_sendspin_json fuzz_sendspin_frames fuzz_sendspin_handshake
+                       fuzz_sendspin_messages)
 
 adm_enabled() { [ -n "${AC3FORGE_FUZZ_ADM:-}" ]; }